app/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
WTForms==2.2.1
pyodbc==4.0.24
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-oauthlib==0.2.0
google-auth-httplib2==0.0.3
requests==2.21.0
